requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.3.1
python-whois==0.8.0
requests==2.31.0
ipinfo==4.4.2
shodan==1.29.1
reportlab==4.0.4
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
beautifulsoup4==4.12.2
urllib3==2.0.7
